+++ Source/Shaders/GlobeVS.glsl
@@ -6,6 +6,14 @@ attribute vec4 position3DAndHeight;
 attribute vec4 textureCoordAndEncodedNormals;
 #endif
 
+#ifdef GEODETIC_SURFACE_NORMALS
+attribute vec3 geodeticSurfaceNormal;
+#endif
+
+#ifdef EXAGGERATION
+uniform vec2 u_terrainExaggerationAndRelativeHeight;
+#endif
+
 uniform vec3 u_center3D;
 uniform mat4 u_modifiedModelView;
 uniform mat4 u_modifiedModelViewProjection;
@@ -156,6 +164,28 @@ void main()
 #endif
 
     vec3 position3DWC = position + u_center3D;
+
+#ifdef GEODETIC_SURFACE_NORMALS
+    vec3 ellipsoidNormal = geodeticSurfaceNormal;
+#else
+    vec3 ellipsoidNormal = normalize(position3DWC);
+#endif
+
+#if defined(EXAGGERATION) && defined(GEODETIC_SURFACE_NORMALS)
+    float exaggeration = u_terrainExaggerationAndRelativeHeight.x;
+    float relativeHeight = u_terrainExaggerationAndRelativeHeight.y;
+    float newHeight = (height - relativeHeight) * exaggeration + relativeHeight;
+
+    // stop from going through center of earth
+    float minRadius = min(min(czm_ellipsoidRadii.x, czm_ellipsoidRadii.y), czm_ellipsoidRadii.z);
+    newHeight = max(newHeight, -minRadius);
+
+    vec3 offset = ellipsoidNormal * (newHeight - height);
+    position += offset;
+    position3DWC += offset;
+    height = newHeight;
+#endif
+
     gl_Position = getPosition(position, height, textureCoordinates);
 
     v_textureCoordinates = vec3(textureCoordinates, webMercatorT);
@@ -164,6 +194,13 @@ void main()
     v_positionEC = (u_modifiedModelView * vec4(position, 1.0)).xyz;
     v_positionMC = position3DWC;  // position in model coordinates
     vec3 normalMC = czm_octDecode(encodedNormal);
+
+#if defined(EXAGGERATION) && defined(GEODETIC_SURFACE_NORMALS)
+    vec3 projection = dot(normalMC, ellipsoidNormal) * ellipsoidNormal;
+    vec3 rejection = normalMC - projection;
+    normalMC = normalize(projection + rejection * exaggeration);
+#endif
+
     v_normalMC = normalMC;
     v_normalEC = czm_normal3D * v_normalMC;
 #elif defined(SHOW_REFLECTIVE_OCEAN) || defined(ENABLE_DAYNIGHT_SHADING) || defined(GENERATE_POSITION) || defined(HDR)
@@ -184,7 +221,6 @@ void main()
 #ifdef APPLY_MATERIAL
     float northPoleZ = czm_ellipsoidRadii.z;
     vec3 northPolePositionMC = vec3(0.0, 0.0, northPoleZ);
-    vec3 ellipsoidNormal = normalize(v_positionMC); // For a sphere this is correct, but not generally for an ellipsoid.
     vec3 vectorEastMC = normalize(cross(northPolePositionMC - v_positionMC, ellipsoidNormal));
     float dotProd = abs(dot(ellipsoidNormal, v_normalMC));
     v_slope = acos(dotProd);