+++ Specs/Core/TerrainEncodingSpec.js
@@ -2,8 +2,11 @@ import { AttributeCompression } from "../../Source/Cesium.js";
 import { AxisAlignedBoundingBox } from "../../Source/Cesium.js";
 import { Cartesian2 } from "../../Source/Cesium.js";
 import { Cartesian3 } from "../../Source/Cesium.js";
+import { Ellipsoid } from "../../Source/Cesium.js";
+import { Math as CesiumMath } from "../../Source/Cesium.js";
 import { Matrix4 } from "../../Source/Cesium.js";
 import { TerrainEncoding } from "../../Source/Cesium.js";
+import { TerrainExaggeration } from "../../Source/Cesium.js";
 import { TerrainQuantization } from "../../Source/Cesium.js";
 import { Transforms } from "../../Source/Cesium.js";
 
@@ -46,6 +49,7 @@ describe("Core/TerrainEncoding", function () {
     var minimumHeight = maximumHeight;
     var hasVertexNormals = false;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -74,6 +78,7 @@ describe("Core/TerrainEncoding", function () {
     var maximumHeight = 100.0;
     var hasVertexNormals = false;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -102,6 +107,7 @@ describe("Core/TerrainEncoding", function () {
     var minimumHeight = maximumHeight;
     var hasVertexNormals = false;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -115,8 +121,8 @@ describe("Core/TerrainEncoding", function () {
     var buffer = [];
     encoding.encode(buffer, 0, position, Cartesian2.ZERO, 100.0);
 
-    expect(encoding.getStride()).toEqual(6);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(6);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodePosition(buffer, 0)).toEqual(position);
   });
@@ -129,6 +135,7 @@ describe("Core/TerrainEncoding", function () {
     var minimumHeight = maximumHeight;
     var hasVertexNormals = true;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -143,8 +150,8 @@ describe("Core/TerrainEncoding", function () {
     var buffer = [];
     encoding.encode(buffer, 0, position, Cartesian2.ZERO, 100.0, normal);
 
-    expect(encoding.getStride()).toEqual(7);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(7);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodePosition(buffer, 0)).toEqual(position);
   });
@@ -152,6 +159,7 @@ describe("Core/TerrainEncoding", function () {
   it("encodes position with quantization and without normals", function () {
     var hasVertexNormals = false;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -165,8 +173,8 @@ describe("Core/TerrainEncoding", function () {
     var buffer = [];
     encoding.encode(buffer, 0, position, Cartesian2.ZERO, 100.0);
 
-    expect(encoding.getStride()).toEqual(3);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(3);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodePosition(buffer, 0)).toEqualEpsilon(position, 1.0);
   });
@@ -174,6 +182,7 @@ describe("Core/TerrainEncoding", function () {
   it("encodes position with quantization and normals", function () {
     var hasVertexNormals = true;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -188,15 +197,76 @@ describe("Core/TerrainEncoding", function () {
     var buffer = [];
     encoding.encode(buffer, 0, position, Cartesian2.ZERO, 100.0, normal);
 
-    expect(encoding.getStride()).toEqual(4);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(4);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodePosition(buffer, 0)).toEqualEpsilon(position, 1.0);
+  });
+
+  it("encodes position without quantization and with exaggeration", function () {
+    var hasVertexNormals = false;
+    var hasWebMercatorT = false;
+    var hasGeodeticSurfaceNormals = true;
+
+    var height = 1000000.0;
+    var position = new Cartesian3(height, 0.0, 0.0);
+    var geodeticSurfaceNormal = new Cartesian3(1.0, 0.0, 0.0);
+
+    var exaggeration = 2.0;
+    var exaggerationRelativeHeight = 10.0;
+    var exaggeratedHeight = TerrainExaggeration.getHeight(
+      height,
+      exaggeration,
+      exaggerationRelativeHeight
+    );
+    var exaggeratedPosition = new Cartesian3(exaggeratedHeight, 0.0, 0.0);
+
+    var maximumHeight = height;
+    var minimumHeight = -height;
+    var maximum = new Cartesian3(height, height, height);
+    var minimum = Cartesian3.negate(maximum, new Cartesian3());
+    var aabox = new AxisAlignedBoundingBox(minimum, maximum, Cartesian3.ZERO);
+
+    var encoding = new TerrainEncoding(
+      aabox.center,
+      aabox,
+      minimumHeight,
+      maximumHeight,
+      Matrix4.IDENTITY,
+      hasVertexNormals,
+      hasWebMercatorT,
+      hasGeodeticSurfaceNormals,
+      exaggeration,
+      exaggerationRelativeHeight
+    );
+
+    var buffer = [];
+    encoding.encode(
+      buffer,
+      0,
+      position,
+      Cartesian2.ZERO,
+      height,
+      undefined,
+      undefined,
+      geodeticSurfaceNormal
+    );
+
+    expect(encoding.stride).toEqual(9);
+    expect(buffer.length).toEqual(encoding.stride);
+    expect(encoding.getExaggeratedPosition(buffer, 0)).toEqualEpsilon(
+      exaggeratedPosition,
+      CesiumMath.EPSILON5
+    );
+    expect(
+      encoding.decodeGeodeticSurfaceNormal(buffer, 0, new Cartesian3())
+    ).toEqualEpsilon(geodeticSurfaceNormal, CesiumMath.EPSILON5);
   });
 
   it("encodes texture coordinates with quantization and without normals", function () {
     var hasVertexNormals = false;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -209,8 +279,8 @@ describe("Core/TerrainEncoding", function () {
     var buffer = [];
     encoding.encode(buffer, 0, Cartesian3.ZERO, texCoords, 100.0);
 
-    expect(encoding.getStride()).toEqual(3);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(3);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodeTextureCoordinates(buffer, 0)).toEqualEpsilon(
       texCoords,
@@ -221,6 +291,7 @@ describe("Core/TerrainEncoding", function () {
   it("encodes textureCoordinates with quantization and normals", function () {
     var hasVertexNormals = true;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -240,8 +311,8 @@ describe("Core/TerrainEncoding", function () {
       Cartesian3.UNIT_X
     );
 
-    expect(encoding.getStride()).toEqual(4);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(4);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodeTextureCoordinates(buffer, 0)).toEqualEpsilon(
       texCoords,
@@ -254,6 +325,7 @@ describe("Core/TerrainEncoding", function () {
     minimumHeight = 0.0;
     maximumHeight = 200.0;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -265,8 +337,8 @@ describe("Core/TerrainEncoding", function () {
     var height = (maximumHeight + minimumHeight) * 0.5;
     encoding.encode(buffer, 0, center, Cartesian2.ZERO, height);
 
-    expect(encoding.getStride()).toEqual(3);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(3);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodeHeight(buffer, 0)).toEqualEpsilon(
       height,
@@ -279,6 +351,7 @@ describe("Core/TerrainEncoding", function () {
     minimumHeight = 0.0;
     maximumHeight = 200.0;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -297,8 +370,8 @@ describe("Core/TerrainEncoding", function () {
       Cartesian3.UNIT_X
     );
 
-    expect(encoding.getStride()).toEqual(4);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(4);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.decodeHeight(buffer, 0)).toEqualEpsilon(
       height,
@@ -309,6 +382,7 @@ describe("Core/TerrainEncoding", function () {
   it("gets oct-encoded normal", function () {
     var hasVertexNormals = true;
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -330,10 +404,75 @@ describe("Core/TerrainEncoding", function () {
       octNormal
     );
 
-    expect(encoding.getStride()).toEqual(4);
-    expect(buffer.length).toEqual(encoding.getStride());
+    expect(encoding.stride).toEqual(4);
+    expect(buffer.length).toEqual(encoding.stride);
 
     expect(encoding.getOctEncodedNormal(buffer, 0)).toEqual(octNormal);
+  });
+
+  it("adds geodetic surface normals", function () {
+    var hasVertexNormals = false;
+    var encoding = new TerrainEncoding(
+      aabox.center,
+      aabox,
+      minimumHeight,
+      maximumHeight,
+      fromENU,
+      hasVertexNormals
+    );
+
+    var oldBuffer = [];
+    encoding.encode(oldBuffer, 0, center, Cartesian2.ZERO, minimumHeight);
+    var oldStride = encoding.stride;
+
+    var newBuffer = [];
+    var ellipsoid = Ellipsoid.UNIT_SPHERE;
+    encoding.addGeodeticSurfaceNormals(oldBuffer, newBuffer, ellipsoid);
+    var newStride = encoding.stride;
+    var strideDifference = newStride - oldStride;
+
+    expect(strideDifference).toEqual(3);
+    expect(oldBuffer.length).toEqual(oldStride);
+    expect(newBuffer.length).toEqual(newStride);
+  });
+
+  it("removes geodetic surface normals", function () {
+    var hasVertexNormals = false;
+    var hasWebMarcatorT = false;
+    var hasGeodeticSurfaceNormals = true;
+    var encoding = new TerrainEncoding(
+      aabox.center,
+      aabox,
+      minimumHeight,
+      maximumHeight,
+      fromENU,
+      hasVertexNormals,
+      hasWebMarcatorT,
+      hasGeodeticSurfaceNormals
+    );
+
+    var geodeticSurfaceNormal = new Cartesian3(1.0, 0.0, 0.0);
+    var oldBuffer = [];
+    encoding.encode(
+      oldBuffer,
+      0,
+      center,
+      Cartesian2.ZERO,
+      minimumHeight,
+      undefined,
+      undefined,
+      geodeticSurfaceNormal
+    );
+    var oldStride = encoding.stride;
+
+    var newBuffer = [];
+    encoding.removeGeodeticSurfaceNormals(oldBuffer, newBuffer);
+    var newStride = encoding.stride;
+    var strideDifference = newStride - oldStride;
+
+    expect(strideDifference).toEqual(-3);
+    expect(oldBuffer.length).toEqual(oldStride);
+    expect(newBuffer.length).toEqual(newStride);
   });
 
   it("gets attributes", function () {
@@ -350,6 +489,7 @@ describe("Core/TerrainEncoding", function () {
     var hasVertexNormals = false;
 
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -378,6 +518,7 @@ describe("Core/TerrainEncoding", function () {
     var hasVertexNormals = false;
 
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -403,6 +544,7 @@ describe("Core/TerrainEncoding", function () {
     var hasVertexNormals = false;
 
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,
@@ -435,6 +577,7 @@ describe("Core/TerrainEncoding", function () {
     var hasVertexNormals = false;
 
     var encoding = new TerrainEncoding(
+      aabox.center,
       aabox,
       minimumHeight,
       maximumHeight,